Pythons/requirements.txt
cmake==3.24.1
dlib==19.22.99
face_recognition==1.3.0
mysql_connector_repackaged==0.3.1
numpy==1.23.2
opencv_python==4.6.0.66
pyserial==3.5
